:truck: Move custom exception from `utility_functions` to `exceptions`
