+++ python_jobs/src/hbnmigration/exceptions/__init__.py
@@ -1,7 +1,10 @@
 """Custom exceptions."""
 
+from .logging import TSVLoggedError
+
 
 class NoData(Exception):  # noqa: N818
     """No data to process."""
 
-    ...
+
+__all__ = ["NoData", "TSVLoggedError"]
