+++ python_jobs/src/hbnmigration/utility_functions/__init__.py
@@ -43,7 +43,7 @@ from .datatypes import (
     ValueClass,
     ValueField,
 )
-from .logging import initialize_logging, setup_tsv_logger, TSVLoggedError
+from .logging import initialize_logging, setup_tsv_logger
 from .secrets import ImportWithFallback
 from .typescript import tsx
 
@@ -65,7 +65,6 @@ __all__ = [
     "InstrumentRowCount",
     "ProjectStatus",
     "Results",
-    "TSVLoggedError",
     "Tokens",
     "ValueClass",
     "ValueField",
