+++ python_jobs/src/hbnmigration/utility_functions/logging.py
@@ -187,20 +187,3 @@ class TSVHandler(logging.Handler):
                 )
         except Exception:
             self.handleError(record)
-
-
-class TSVLoggedError(Exception):
-    """Exception that logs to TSV."""
-
-    logger = setup_tsv_logger("mrn_error_log", "mrn_error_log.tsv")
-
-    def __init__(self, mrn: int | str, error_message: Optional[str], attempt: str):
-        """Initialize TSV logging exception."""
-        self.mrn = mrn
-        self.error_message = error_message or ""
-        self.attempt = attempt
-
-        # Log using logging library
-        self.logger.error(error_message, extra={"mrn": mrn, "attempt": attempt})
-
-        super().__init__(f"MRN {mrn}: {error_message} (Attempt {attempt})")
